Revert "breaking: Use two modules to handle cjs / esmodule export differences"
This reverts commit cc9ac56bc27c7c601e209f1bdae1210dd06802b2.
