--- src/es.ts
@@ -1,3 +1,0 @@
-import { Parser, protractorFlake } from './protractor-flake'
-
-export { Parser, protractorFlake }
